Correction de lect audio -Il y'avait une erreur dans la lecture des données il manquait une valeur absolue dans 20log(fab(value)) et donc les amplitudes lues n'étaient pas bonne ! -> c'est corrigé et vérifier en comparant à la lecture en matlab qui donnes les mêmes valeurs d'amplitudes (à une translation prêt mais pas gênant) -Les données d'amplitudes sont désormais stockés comme des shorts afin d'être plus rapide -La fonction piano_audio_to_piano_note tourne à une vitesse raisonnable cependant les fréquence prépondérente ne sont pas les bonnes (il va falloir chercher les erreur en comparant au code Matlab)

+++ src/fenetrage_hamming.c
@@ -4,10 +4,10 @@
 #include <math.h>
 
 
-void fenetrage_hamming(float* tab_amplitude, float* amplitude_fenetree, int nb_data, float Fe, float t1, float t2)
+void fenetrage_hamming(short* tab_amplitude, short* amplitude_fenetree, int nb_data, float Fe, float t1, float t2)
 {
     float Te = 1/Fe;
-    for(int k=0;k<nb_data;k++)
+    for(int k=0; k < nb_data; k++)
     {
         float t = k*Te;
         if(t>t1 && t<t2)
